requirements.txt
bson==0.5.8
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.7
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.41.0
django-cors-headers==3.2.1
django-rest-auth==0.9.5
django-rest-framework-mongoengine==3.4.1
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djongo==1.3.0
gunicorn==20.0.4
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mongoengine==0.19.1
oauthlib==3.1.0
openapi-codec==1.3.2
psycopg2==2.8.4
pymongo==3.10.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.2.4
uritemplate==3.0.1
urllib3==1.25.7
whitenoise==5.0.1
